Separate Controller for NewCompanyInsertion Created!
Code is moved from HomeController to NewCompanyController... Required
changes are made in routes.php
